Moved tests to new ai_test_test_runner.c file
